# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * AutoImport CMS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthCredentials {
  username: string;
  password: string;
}

export interface AuthUser {
  id: number;
  username: string;
  role: string;
}

export interface Vehicle {
  id: number;
  make: string;
  model: string;
  year: number;
  engine: string;
  fuel: string;
  transmission: string;
  mileage: number;
  location: string;
  price: number;
  description: string;
  status: string;
  imageUrl: string;
  /** @nullable */
  badge?: string | null;
  isPopular?: boolean;
  sortOrder: number;
  createdAt?: string;
  updatedAt?: string;
}

export interface VehicleInput {
  make: string;
  model: string;
  year: number;
  engine: string;
  fuel: string;
  transmission: string;
  mileage: number;
  location: string;
  price: number;
  description: string;
  status?: string;
  imageUrl?: string;
  /** @nullable */
  badge?: string | null;
  isPopular?: boolean;
  sortOrder?: number;
}

export interface VehicleUpdate {
  make?: string;
  model?: string;
  year?: number;
  engine?: string;
  fuel?: string;
  transmission?: string;
  mileage?: number;
  location?: string;
  price?: number;
  description?: string;
  status?: string;
  imageUrl?: string;
  /** @nullable */
  badge?: string | null;
  isPopular?: boolean;
  sortOrder?: number;
}

export interface SoldVehicle {
  id: number;
  make: string;
  model: string;
  year: number;
  /** @nullable */
  finalPrice?: number | null;
  purchaseCountry: string;
  deliveryStatus: string;
  /** @nullable */
  deliveryDate?: string | null;
  imageUrl: string;
  createdAt?: string;
}

export interface SoldVehicleInput {
  make: string;
  model: string;
  year: number;
  /** @nullable */
  finalPrice?: number | null;
  purchaseCountry: string;
  deliveryStatus?: string;
  /** @nullable */
  deliveryDate?: string | null;
  imageUrl?: string;
}

export interface SoldVehicleUpdate {
  make?: string;
  model?: string;
  year?: number;
  /** @nullable */
  finalPrice?: number | null;
  purchaseCountry?: string;
  deliveryStatus?: string;
  /** @nullable */
  deliveryDate?: string | null;
  imageUrl?: string;
}

export interface PopularVehicle {
  id: number;
  make: string;
  model: string;
  imageUrl: string;
  priceRange: string;
  estimatedDelivery: string;
  description: string;
  sortOrder: number;
}

export interface PopularVehicleInput {
  make: string;
  model: string;
  imageUrl?: string;
  priceRange: string;
  estimatedDelivery: string;
  description: string;
  sortOrder?: number;
}

export interface PopularVehicleUpdate {
  make?: string;
  model?: string;
  imageUrl?: string;
  priceRange?: string;
  estimatedDelivery?: string;
  description?: string;
  sortOrder?: number;
}

export interface Service {
  id: number;
  iconName: string;
  titleEn: string;
  titlePl?: string;
  titleRu?: string;
  descriptionEn: string;
  descriptionPl?: string;
  descriptionRu?: string;
  sortOrder: number;
  createdAt?: string;
}

export interface ServiceInput {
  iconName?: string;
  titleEn: string;
  titlePl?: string;
  titleRu?: string;
  descriptionEn: string;
  descriptionPl?: string;
  descriptionRu?: string;
  sortOrder?: number;
}

export interface ServiceUpdate {
  iconName?: string;
  titleEn?: string;
  titlePl?: string;
  titleRu?: string;
  descriptionEn?: string;
  descriptionPl?: string;
  descriptionRu?: string;
  sortOrder?: number;
}

export interface PricingPackage {
  id: number;
  nameEn: string;
  namePl?: string;
  nameRu?: string;
  price: number;
  currency: string;
  features: string[];
  isPopular: boolean;
  sortOrder: number;
}

export interface PricingPackageInput {
  nameEn: string;
  namePl?: string;
  nameRu?: string;
  price: number;
  currency?: string;
  features?: string[];
  isPopular?: boolean;
  sortOrder?: number;
}

export interface PricingPackageUpdate {
  nameEn?: string;
  namePl?: string;
  nameRu?: string;
  price?: number;
  currency?: string;
  features?: string[];
  isPopular?: boolean;
  sortOrder?: number;
}

export interface Testimonial {
  id: number;
  name: string;
  country: string;
  vehicleName: string;
  rating: number;
  content: string;
  /** @nullable */
  avatarUrl?: string | null;
  isActive: boolean;
  createdAt?: string;
}

export interface TestimonialInput {
  name: string;
  country: string;
  vehicleName: string;
  rating?: number;
  content: string;
  /** @nullable */
  avatarUrl?: string | null;
  isActive?: boolean;
}

export interface TestimonialUpdate {
  name?: string;
  country?: string;
  vehicleName?: string;
  rating?: number;
  content?: string;
  /** @nullable */
  avatarUrl?: string | null;
  isActive?: boolean;
}

export interface PageContent {
  id: number;
  page: string;
  sectionKey: string;
  valueEn: string;
  valuePl: string;
  valueRu: string;
  updatedAt?: string;
}

export interface PageContentUpdate {
  valueEn?: string;
  valuePl?: string;
  valueRu?: string;
}

export interface SiteSetting {
  id: number;
  key: string;
  value: string;
  label?: string;
  group?: string;
  updatedAt?: string;
}

export interface SiteSettingUpdate {
  value: string;
}