# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AutoImport CMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vehicles:
    get:
      operationId: getVehicles
      tags: [vehicles]
      responses:
        "200":
          description: List of vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
    post:
      operationId: createVehicle
      tags: [vehicles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
  /vehicles/{id}:
    get:
      operationId: getVehicle
      tags: [vehicles]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "404":
          description: Not found
    put:
      operationId: updateVehicle
      tags: [vehicles]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
    delete:
      operationId: deleteVehicle
      tags: [vehicles]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Vehicle:
      type: object
      required: [id, make, model, year, engine, fuel, transmission, mileage, location, price, description, status, imageUrl, sortOrder]
      properties:
        id:
          type: integer
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        engine:
          type: string
        fuel:
          type: string
        transmission:
          type: string
        mileage:
          type: integer
        location:
          type: string
        price:
          type: integer
        description:
          type: string
        status:
          type: string
        imageUrl:
          type: string
        badge:
          type: ["string", "null"]
        isPopular:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    VehicleInput:
      type: object
      required: [make, model, year, engine, fuel, transmission, mileage, location, price, description]
      properties:
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        engine:
          type: string
        fuel:
          type: string
        transmission:
          type: string
        mileage:
          type: integer
        location:
          type: string
        price:
          type: integer
        description:
          type: string
        status:
          type: string
        imageUrl:
          type: string
        badge:
          type: ["string", "null"]
        isPopular:
          type: boolean
        sortOrder:
          type: integer
    VehicleUpdate:
      type: object
      properties:
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        engine:
          type: string
        fuel:
          type: string
        transmission:
          type: string
        mileage:
          type: integer
        location:
          type: string
        price:
          type: integer
        description:
          type: string
        status:
          type: string
        imageUrl:
          type: string
        badge:
          type: ["string", "null"]
        isPopular:
          type: boolean
        sortOrder:
          type: integer
...