# lib/api-zod/src/generated/types/vehicle.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * AutoImport CMS API
 * OpenAPI spec version: 0.1.0
 */

export interface Vehicle {
  id: number;
  make: string;
  model: string;
  year: number;
  engine: string;
  fuel: string;
  transmission: string;
  mileage: number;
  location: string;
  price: number;
  description: string;
  status: string;
  imageUrl: string;
  /** @nullable */
  badge?: string | null;
  isPopular?: boolean;
  sortOrder: number;
  createdAt?: string;
  updatedAt?: string;
}