# lib/api-zod/src/generated/types/vehicleUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * AutoImport CMS API
 * OpenAPI spec version: 0.1.0
 */

export interface VehicleUpdate {
  make?: string;
  model?: string;
  year?: number;
  engine?: string;
  fuel?: string;
  transmission?: string;
  mileage?: number;
  location?: string;
  price?: number;
  description?: string;
  status?: string;
  imageUrl?: string;
  /** @nullable */
  badge?: string | null;
  isPopular?: boolean;
  sortOrder?: number;
}